requirements.txt
Flask==1.0.2
line-bot-sdk==1.8.0
post==2018.11.20
request==2018.11.20
requests==2.21.0
janome==0.3.7